2018.7.25 20:41 订单接口的测试以及权限系统的初步开发

+++ src/main/java/com/amazon/module/entity/ProductCategory.java
@@ -1,15 +1,18 @@
 package com.amazon.module.entity;
 
+/**
+ * @describe 商品分类实体类
+ * */
 public class ProductCategory {
-    private long pc_id;
+    private int pc_id;
     private String name;
-    private long pc_parent_id;
+    private int pc_parent_id;
 
-    public long getPc_id() {
+    public int getPc_id() {
         return pc_id;
     }
 
-    public void setPc_id(long pc_id) {
+    public void setPc_id(int pc_id) {
         this.pc_id = pc_id;
     }
 
@@ -21,11 +24,11 @@ public class ProductCategory {
         this.name = name;
     }
 
-    public long getPc_parent_id() {
+    public int getPc_parent_id() {
         return pc_parent_id;
     }
 
-    public void setPc_parent_id(long pc_parent_id) {
+    public void setPc_parent_id(int pc_parent_id) {
         this.pc_parent_id = pc_parent_id;
     }
 }
